config moved to utils
